requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
cryptography==3.4.7
decorator==5.0.9
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-core==1.7.2
google-cloud-translate==3.3.2
googleapis-common-protos==1.53.0
grpcio==1.38.1
idna==3.1
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
multidict==5.1.0
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pip==21.2.3
prompt-toolkit==3.0.19
proto-plus==1.19.0
protobuf==3.17.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
pytz==2021.1
pyu2f==0.1.5
requests==2.26.0
rsa==4.7.2
setuptools==49.6.0.post20210108
six==1.16.0
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
wheel==0.36.2
yarl==1.6.3
